+++ packages/next-intl/src/navigation/react-client/useBaseRouter.tsx
@@ -2,9 +2,9 @@ import {useRouter as useNextRouter, usePathname} from 'next/navigation';
 import {useMemo} from 'react';
 import useLocale from '../../react-client/useLocale';
 import {Locales, LocalePrefixConfigVerbose} from '../../routing/types';
-import {getLocalePrefix, localizeHref} from '../../shared/utils';
+import {getLocalePrefix} from '../../shared/utils';
 import syncLocaleCookie from '../shared/syncLocaleCookie';
-import {getBasePath} from '../shared/utils';
+import {getBasePath, localizeHref} from '../shared/utils';
 
 type IntlNavigateOptions<AppLocales extends Locales> = {
   locale?: AppLocales[number];
