+++ packages/next-intl/test/shared/utils.test.tsx
@@ -1,10 +1,5 @@
 import {it, describe, expect} from 'vitest';
-import {
-  hasPathnamePrefixed,
-  unprefixPathname,
-  matchesPathname,
-  prefixPathname
-} from '../../src/shared/utils';
+import {matchesPathname, prefixPathname} from '../../src/shared/utils';
 
 describe('prefixPathname', () => {
   it("doesn't add trailing slashes for the root", () => {
@@ -17,41 +12,6 @@ describe('prefixPathname', () => {
 
   it('localizes nested paths', () => {
     expect(prefixPathname('/en', '/nested')).toEqual('/en/nested');
-  });
-});
-
-describe('hasPathnamePrefixed', () => {
-  it('detects prefixed pathnames', () => {
-    expect(hasPathnamePrefixed('/en', '/en')).toEqual(true);
-    expect(hasPathnamePrefixed('/en', '/en/')).toEqual(true);
-    expect(hasPathnamePrefixed('/en', '/en/client')).toEqual(true);
-    expect(hasPathnamePrefixed('/en', '/en/client/')).toEqual(true);
-    expect(hasPathnamePrefixed('/en', '/en/client/test')).toEqual(true);
-  });
-
-  it('detects non-prefixed pathnames', () => {
-    expect(hasPathnamePrefixed('/en', '/')).toEqual(false);
-    expect(hasPathnamePrefixed('/en', '/client')).toEqual(false);
-    expect(hasPathnamePrefixed('/en', '/client/')).toEqual(false);
-    expect(hasPathnamePrefixed('/en', '/client/test')).toEqual(false);
-  });
-});
-
-describe('unlocalizePathname', () => {
-  it('works for the root', () => {
-    expect(unprefixPathname('/en', '/en')).toEqual('/');
-  });
-
-  it('works for nested pages', () => {
-    expect(unprefixPathname('/en/nested', '/en')).toEqual('/nested');
-  });
-
-  it('works with sub-tags', () => {
-    expect(unprefixPathname('/en-GB/nested', '/en-GB')).toEqual('/nested');
-  });
-
-  it('works for custom prefixes', () => {
-    expect(unprefixPathname('/uk/nested', '/uk')).toEqual('/nested');
   });
 });
